requirements.txt
Django==2.0.2
django-webpack-loader==0.6.0
djangorestframework==3.7.7
ipython==6.2.1
requests-2.18.4
